revert and cleanup

+++ addons/ofxGui/src/ofxInputField.cpp
@@ -273,7 +273,7 @@ void ofxInputField<Type>::moveCursor(int cursorPos){
 	selectStartPos = selectEndPos = cursorPos;
 	selectionWidth = 0;
 	
-	if ( visibleInputEnd < (int)ofUTF8Length( input ) ){
+	if ( visibleInputEnd < ofUTF8Length( input ) ){
 		// Attempt to extend visible input range. If a character was being entered 
 		// somewhere inside the text, there might be enough space to the right to 
 		// show the newly extended input.
